requirements.txt
attrs==22.1.0
autotrader @ git+https://github.com/thelord810/AutoTrader.git@f8085154a128ac19e7e58cb8f26d9fb88a00527d
bidict==0.22.0
bokeh==2.4.3
breeze-connect==1.0.20
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
confluent-kafka==1.9.2
eventkit==0.8.9
finta==1.3
ib-insync==0.9.71
idna==3.3
importlib-resources==5.9.0
Jinja2==3.1.2
jsonschema==4.15.0
lxml==4.9.1
MarkupSafe==2.1.1
multitasking==0.0.11
nest-asyncio==1.5.5
numpy==1.23.2
packaging==21.3
pandas==1.4.4
pendulum==2.1.2
Pillow==9.2.0
pyfiglet==0.8.post1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scipy==1.9.1
six==1.16.0
tornado==6.2
tqdm==4.64.1
typing-extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
v20==3.0.25.0
websocket-client==1.4.0
yfinance==0.1.74

numpy >= 1.20.3
pandas >= 1.3.4
art >= 5.7
PyYAML
bokeh >= 2.3.1
scipy >= 1.7.1
finta >= 1.3
tqdm>=4.64.0
importlib-resources
click >= 8.1.3
requests >= 2.28.1
python-telegram-bot >= 13.14
